graph: rename directory from graphsession to session
The package name is kept the same. So this is just for less stuttering
at a path level.
