+++ graph/session/graph_session.go
@@ -1,4 +1,4 @@
-package graphsession
+package session
 
 import (
 	"fmt"
